Add : List 컴포넌트에서 Lmap을 돌려  ListProduct 컴포넌트에 props 전달완하고 서버 확인까지 완료했음
