--- src/pages/List/List.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-
-const List = () => {
-  return <Text>리스트!</Text>;
-};
-
-const Text = styled.p``;
-
-export default List;
